+++ stars-core/src/test/kotlin/tools/aqua/stars/core/SimpleTypes.kt
@@ -40,6 +40,7 @@ class SimpleEntity(
 class SimpleTickData(
     currentTickUnit: SimpleTickDataUnit = SimpleTickDataUnit(0),
     entities: Set<SimpleEntity> = LinkedHashSet(),
+    override val identifier: String = "SimpleTickData",
 ) :
     TickDataType<SimpleEntity, SimpleTickData, SimpleTickDataUnit, SimpleTickDataDifference>(
         currentTickUnit,