+++ stars-importer-carla/src/main/kotlin/tools/aqua/stars/importer/carla/carlaDataFileHandling.kt
@@ -149,6 +149,7 @@ fun loadTicks(
           convertTickData(
               world = currentSimulationRunsWrapper.world,
               jsonSimulationRun = simulationRun,
+              tickDataSourcePath = currentDynamicDataPath,
           )
 
       val iterator = ticks.iterator()